refactor: split AIChatSidebar into smaller components following react-doctor feedback
Extract 8 sub-components from the 1182-line AIChatSidebar:
- IntelligenceInsight, ShimmerText, PostHogFeedback, SidebarHeader,
  SidebarFloatingButton, EmptyState, UserMessageBubble, AIMessageCard,
  ChatInput
- Shared Message/MessageVersion types in ai-chat-types.ts
- AIChatSidebar reduced from 1182 to 467 lines

No functional changes — all existing behavior preserved.
